+++ src/main/kotlin/org/clothfoundation/cloth/editor/ClothQuoteHandler.kt
@@ -1,0 +1,9 @@
+package org.clothfoundation.cloth.editor
+
+import com.intellij.codeInsight.editorActions.SimpleTokenSetQuoteHandler
+import org.clothfoundation.cloth.lexer.ClothTokenTypes
+
+class ClothQuoteHandler : SimpleTokenSetQuoteHandler(
+    ClothTokenTypes.STRING_LITERAL,
+    ClothTokenTypes.CHARACTER_LITERAL,
+)